Rename eosjs2 -> eosjs
