+++ docs/3.-Browsers.md
@@ -3,20 +3,20 @@
 ## Usage
 `npm run build-web` or `yarn build-web`
 
-Reuse the `api` object for all transactions; it caches ABIs to reduce network usage. Only call `new eosjs2.Api(...)` once.
+Reuse the `api` object for all transactions; it caches ABIs to reduce network usage. Only call `new eosjs.Api(...)` once.
 
 ```html
 <pre style="width: 100%; height: 100%; margin:0px; "></pre>
 
-<script src='dist-web/eosjs2-debug.js'></script>
-<script src='dist-web/eosjs2-jsonrpc-debug.js'></script>
-<script src='dist-web/eosjs2-jssig-debug.js'></script>
+<script src='dist-web/eosjs-debug.js'></script>
+<script src='dist-web/eosjs-jsonrpc-debug.js'></script>
+<script src='dist-web/eosjs-jssig-debug.js'></script>
 <script>
   let pre = document.getElementsByTagName('pre')[0];
   const defaultPrivateKey = "5JtUScZK2XEp3g9gh7F8bwtPTRAkASmNrrftmx4AxDKD5K4zDnr"; // useraaaaaaaa
-  const rpc = new eosjs2_jsonrpc.JsonRpc('http://127.0.0.1:8000');
-  const signatureProvider = new eosjs2_jssig.default([defaultPrivateKey]);
-  const api = new eosjs2.Api({ rpc, signatureProvider });
+  const rpc = new eosjs_jsonrpc.JsonRpc('http://127.0.0.1:8000');
+  const signatureProvider = new eosjs_jssig.default([defaultPrivateKey]);
+  const api = new eosjs.Api({ rpc, signatureProvider });
 
   (async () => {
     try {
@@ -42,7 +42,7 @@ Reuse the `api` object for all transactions; it caches ABIs to reduce network us
       pre.textContent += '\n\nTransaction pushed!\n\n' + JSON.stringify(result, null, 2);
     } catch (e) {
       pre.textContent = '\nCaught exception: ' + e;
-      if (e instanceof eosjs2_jsonrpc.RpcError)
+      if (e instanceof eosjs_jsonrpc.RpcError)
         pre.textContent += '\n\n' + JSON.stringify(e.json, null, 2);
     }
   })();
@@ -50,4 +50,4 @@ Reuse the `api` object for all transactions; it caches ABIs to reduce network us
 ```
 
 ## IE11 and Edge Support
-If you need to support IE11 or Edge you will also need to install a text-encoding polyfill as eosjs2 Signing is dependent on the TextEncoder which IE11 and Edge do not provide.  Pass the TextEncoder and TextDecoder to the API constructor as demonstrated in the [ES 2015 example](#node-es-2015).  Refer to the documentation here https://github.com/inexorabletash/text-encoding to determine the best way to include it in your project.
+If you need to support IE11 or Edge you will also need to install a text-encoding polyfill as eosjs Signing is dependent on the TextEncoder which IE11 and Edge do not provide.  Pass the TextEncoder and TextDecoder to the API constructor as demonstrated in the [ES 2015 example](#node-es-2015).  Refer to the documentation here https://github.com/inexorabletash/text-encoding to determine the best way to include it in your project.
