+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "eosjs2",
+  "name": "eosjs",
   "version": "1.0.0",
   "description": "Talk to eos API",
   "main": "dist/index.js",
@@ -20,7 +20,7 @@
   "license": "MIT",
   "repository": {
     "type": "git",
-    "url": "https://github.com/EOSIO/eosjs2.git"
+    "url": "https://github.com/EOSIO/eosjs.git"
   },
   "dependencies": {
     "@types/jest": "^23.3.1",
